JWT agregados para autenticación + Inicio de funcionalidad para manejo de amigos

+++ database/client.js
@@ -1,11 +1,9 @@
 const { MongoClient, ObjectId } = require('mongodb');
 
-const uri = `mongodb+srv://${process.env.MONGO_USER}:${process.env.MONGO_PASS}@wooks.swarj.mongodb.net/${process.env.MONGO_DB}?retryWrites=true&w=majority`;
-
 class MongoDB {
     constructor() {
-        this.client = new MongoClient(uri, {
-             useNewUrlParser: true, useUnifiedTopology: true 
+        this.client = new MongoClient(process.env.MONGO_URI, {
+            useNewUrlParser: true, useUnifiedTopology: true
         });
 
         this.dbname = process.env.MONGO_DB;
@@ -22,7 +20,7 @@ class MongoDB {
 
                         console.log("BASE DE DATOS CONECTADA");
 
-                        resolve(this.client.db(this.dbname).collection('books'));
+                        resolve(this.client.db(this.dbname));
                     })
                 }
             )
@@ -31,63 +29,65 @@ class MongoDB {
         return this.connection;
     }
 
-    getAll() {
+    getAll(userId) {
         return this.connect().then((db) => {
-            return db.find().toArray();
+            return db.collection('books').find({id_owner: ObjectId(userId)}).toArray();
         })
     }
 
-    getOne(id) {
+    getOne(id, userId) {
         return this.connect().then((db) => {
-            return db.find({"_id": ObjectId(id)}).toArray();
+            return db.collection('books').find({ "_id": ObjectId(id), id_owner: ObjectId(userId) }).toArray();
         })
     }
 
-    getCantidadPaginas (id) {
+    getCantidadPaginas(id, userId) {
         return this.connect().then((db) => {
-            return db.find({"_id": ObjectId(id)}).project({"_id": 0, "paginas": 1}).toArray();
+            return db.collection('books').find({ "_id": ObjectId(id), id_owner: ObjectId(userId) }).project({ "_id": 0, "paginas": 1 }).toArray();
         })
     }
 
-    create(newBook) {
+    create(newBook, userId) {
         return this.connect().then((db) => {
-            return db.insertOne(newBook);
+            return db.collection('books').insertOne({...newBook, id_owner: ObjectId(userId)});
         })
     }
 
-    delete(id) {
+    delete(id, userId) {
         return this.connect().then((db) => {
-            return db.deleteOne({"_id": ObjectId(id)});
+            return db.collection('books').deleteOne({ "_id": ObjectId(id), id_owner: ObjectId(userId) });
         })
     }
 
-    updateStatus(id, payload) {
-        
+    updateStatus(id, payload, userId) {
+
         return this.connect().then((db) => {
             if (payload.NewState === "Terminado") {
-                return db.updateOne({"_id": ObjectId(id)}, 
-                    {"$set": {"state": payload.NewState, "review": parseInt(payload.review)},
-                    "$unset": {"actual": "", "progress": ""}});
+                return db.collection('books').updateOne({ "_id": ObjectId(id), id_owner: ObjectId(userId)},
+                    {
+                        "$set": { "state": payload.NewState, "review": parseInt(payload.review) },
+                        "$unset": { "actual": "", "progress": "" }
+                    });
             } else if (payload.NewState === "Leyendo") {
-                return db.updateOne({"_id": ObjectId(id)}, 
-                    {"$set": {"state": payload.NewState, "actual": 1, "progress": 0}});
+                return db.collection('books').updateOne({ "_id": ObjectId(id), id_owner: ObjectId(userId)},
+                    { "$set": { "state": payload.NewState, "actual": 1, "progress": 0 } });
             } else {
-                return db.updateOne({"_id": ObjectId(id)}, 
-                    {"$set": {"state": payload.NewState}});
+                return db.collection('books').updateOne({ "_id": ObjectId(id), id_owner: ObjectId(userId)},
+                    { "$set": { "state": payload.NewState } });
             }
         })
     }
 
-    updateProgress(id, newActual, totalPages){
-        
+    updateProgress(id, newActual, totalPages, userId) {
+
         if (totalPages === 0) {
             return false;
         } else {
 
-            let newProgress = parseInt(newActual/totalPages*100);
+            let newProgress = parseInt(newActual / totalPages * 100);
 
             return this.connect().then((db) => {
-                return db.updateOne({"_id": ObjectId(id)}, {
+                return db.collection('books').updateOne({ "_id": ObjectId(id), id_owner: ObjectId(userId) }, {
                     "$set": {
                         "actual": newActual,
                         "progress": newProgress
@@ -97,7 +97,9 @@ class MongoDB {
         }
     }
 
-    updateBook(id, changes) {
+    updateBook(id, changes, userId) {
+
+        /** MIRAR: HACE FALTA REFACTORIZAR O YA FUNCIONA CORRECTAMENTE? */
 
         let BookChanges = {}
 
@@ -108,9 +110,71 @@ class MongoDB {
         changes.categories ? BookChanges.categories = changes.categories : "";
 
         return this.connect().then((db) => {
-            return db.updateOne({"_id": ObjectId(id)}, {
-                "$set": {...BookChanges}
+            return db.collection('books').updateOne({ "_id": ObjectId(id), id_owner: ObjectId(userId) }, {
+                "$set": { ...BookChanges }
             })
+        })
+    }
+
+    /** USUARIOS */
+    insertUser(data) {
+        return this.connect().then((db) => {
+            return db.collection('users').insertOne(data);
+        });
+    }
+    getUser(userEmail) {
+        return this.connect().then((db) => {
+            return db.collection('users').findOne({ email: userEmail });
+        });
+    }
+    getUserFriends(userId) {
+        return this.connect().then((db) => {
+            return db.collection('users').aggregate([
+                {
+                  '$match': {
+                    '_id': new ObjectId(userId)
+                  }
+                }, {
+                  '$lookup': {
+                    'from': 'users', 
+                    'localField': 'friends', 
+                    'foreignField': '_id', 
+                    'as': 'amigos'
+                  }
+                }
+              ])
+        });
+    }
+    updateUser(userId, editedUser) {
+        return this.connect().then((db) => {
+            return db.collection('users').updateOne(
+                { _id: ObjectId(userId) },
+                { $set: { ...editedUser } }
+            );
+        });
+    }
+    deleteUser(userId) {
+        return this.connect().then((db) => {
+            return db.collection('users').deleteOne({ _id: ObjectId(userId) });
+        });
+    }
+
+    /* AMIGOS */
+    requestFriend(userId, FriendId) {
+        return this.connect().then((db) => {
+            return db.collection('users').updateOne({ _id: ObjectId(FriendId) }, { $push: { request_friends: ObjectId(userId) } })
+        })
+    }
+
+    addFriend(userId, FriendId) {
+        return this.connect().then((db) => {
+
+            // Eliminar la solicitud de amistad
+            db.collection('users').updateOne({ _id: ObjectId(userId) }, { $pull: { request_friends: ObjectId(userId) } })
+
+            // Agregar como amigos a ambos
+                   db.collection('users').updateOne({ _id: ObjectId(userId) }, { $push: { friends: ObjectId(FriendId) } })
+            return db.collection('users').updateOne({ _id: ObjectId(FriendId) }, { $push: { friends: ObjectId(userId) } })
         })
     }
 };
